requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.13.0
pytz==2025.2
Werkzeug==3.1.3
python-decouple==3.8
